Move the generic CLI logic to a common module
This makes way for the use of these common bits in another
CLI module.
