+++ litezip/cli/completezip2litezip.py
@@ -6,7 +6,8 @@ import sys
 from pathlib import Path
 
 from litezip import convert_completezip
-from litezip.logger import configure_logging, logger
+from litezip.logger import logger
+from .common import set_verbosity
 
 
 def _arg_parser():
@@ -29,47 +30,6 @@ def _arg_parser():
         'location',
         help="Location of the unpacked completezip")
     return parser
-
-
-console_logging_config = {
-    'version': 1,
-    'formatters': {
-        'cli': {
-            'format': '%(levelname)-5.5s: %(message)s',
-            },
-    },
-    'filters': {},
-    'handlers': {
-        'console': {
-            'class': 'logging.StreamHandler',
-            'formatter': 'cli',
-            'filters': [],
-            'stream': 'ext://sys.stdout',
-        },
-    },
-    'loggers': {
-        'litezip': {
-            'level': 'ERROR',
-            'handlers': ['console'],
-            'propagate': 0,
-        },
-    },
-    'root': {
-        'level': 'NOTSET',
-        'handlers': [],
-    },
-}
-
-
-def set_verbosity(verbose):
-    config = console_logging_config.copy()
-    if verbose is None:
-        config['loggers']['litezip']['level'] = 'ERROR'
-    elif verbose:
-        config['loggers']['litezip']['level'] = 'DEBUG'
-    else:  # quiet
-        config['loggers']['litezip']['level'] = 100
-    configure_logging(config)
 
 
 def main(argv=None):
